requirements.txt
cv2
sklearn
tensorflow
tensorflow_hub
Pillow
numpy
scipy